Day 25: Running Time and Complexity.py
First, we use some magical math tools that someone created for us called "math" to help us out. We need it for square root calculations.

Then, we have a special function called "isPrime(n)" that will do the magic for us. It's like having a friendly wizard who can tell us if a number is prime or not.

The "def isPrime(n):" part tells the computer that we're going to create a new function called "isPrime." Inside the function, we check if the number "n" is less than or equal to 1. If it is, then it's not a prime number, and we return "False" to say it's not prime.

Next, we use the magical math tool to find the square root of "n" and save it as "sqrt_n." A square root is like finding the number that, when multiplied by itself, gives us "n."

If "sqrt_n" turns out to be a whole number (like 4 or 9), it means "n" is not a prime number, because it has other factors besides 1 and itself. So we return "False" again.

After that, we use a cool loop that goes through all the numbers from 2 to the square root of "n" (rounded up to the nearest whole number). This loop helps us check if there are any other numbers (besides 1 and "n") that can divide "n" evenly. If we find any, it means "n" is not prime, and we return "False" once more.

But if we go through the loop without finding any other numbers that can divide "n," then "n" must be a prime number! So, we return "True" to show it's prime.

Now, let's use our magical "isPrime" function in the main part of the code:

We first ask you to tell us how many numbers you want to check. You give us that number, and we call it "num_test_cases."

Then, we have a loop that runs "num_test_cases" times. Each time it goes through the loop, we ask you to give us a number ("n") to check if it's prime or not.

We use our magical "isPrime" function to check if the number you gave us is prime or not. If it's prime, we print "Prime" to celebrate! 🎉 If it's not prime, we print "Not prime" to let you know.